pyproject.toml
[project]
name = "genai-intern-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.17.0",
    "fastapi[standard]>=0.116.1",
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.6.5",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "ruff>=0.12.8",
    "sentence-transformers>=5.1.0",
    "textstat>=0.7.8",
]
